# PYTHON-3508 Improve the performance of GridOut.readline and GridOut.read