--- a/gridfs/grid_file.py
+++ b/gridfs/grid_file.py
@@ -463,7 +463,10 @@ class GridOut(io.IOBase):
         self.__files = root_collection.files
         self.__file_id = file_id
         self.__buffer = EMPTY
+        # Start position within the current buffered chunk.
+        self.__buffer_pos = 0
         self.__chunk_iter = None
+        # Position within the total file.
         self.__position = 0
         self._file = file_document
         self._session = session
@@ -510,12 +513,12 @@ class GridOut(io.IOBase):
         """Reads a chunk at a time. If the current position is within a
         chunk the remainder of the chunk is returned.
         """
-        received = len(self.__buffer)
+        received = len(self.__buffer) - self.__buffer_pos
         chunk_data = EMPTY
         chunk_size = int(self.chunk_size)
 
         if received > 0:
-            chunk_data = self.__buffer
+            chunk_data = self.__buffer[self.__buffer_pos :]
         elif self.__position < int(self.length):
             chunk_number = int((received + self.__position) / chunk_size)
             if self.__chunk_iter is None:
@@ -531,6 +534,7 @@ class GridOut(io.IOBase):
 
         self.__position += len(chunk_data)
         self.__buffer = EMPTY
+        self.__buffer_pos = 0
         return chunk_data
 
     def _read_size_or_line(self, size: int = -1, line: bool = False) -> bytes:
@@ -545,28 +549,37 @@ class GridOut(io.IOBase):
 
         received = 0
         data = []
-        extra = EMPTY
         while received < size:
             needed = size - received
-            chunk_data = self.readchunk()
-            received += len(chunk_data)
+            if self.__buffer:
+                # Optimization: Read the buffer with zero byte copies.
+                buf = self.__buffer
+                chunk_start = self.__buffer_pos
+                chunk_data = memoryview(buf)[self.__buffer_pos :]
+                self.__buffer = EMPTY
+                self.__buffer_pos = 0
+                self.__position += len(chunk_data)
+            else:
+                buf = chunk_data = self.readchunk()
+                chunk_start = 0
             if line:
-                pos = chunk_data.find(NEWLN, 0, needed)
-                if pos != -1:
-                    data.append(chunk_data[: pos + 1])
-                    extra = chunk_data[pos + 1 :]
-                    break
+                pos = buf.find(NEWLN, chunk_start, chunk_start + needed) - chunk_start
+                if pos >= 0:
+                    # Decrease size to exit the loop.
+                    size = received + pos + 1
+                    needed = pos + 1
             if len(chunk_data) > needed:
                 data.append(chunk_data[:needed])
-                extra = chunk_data[needed:]
+                # Optimization: Save the buffer with zero byte copies.
+                self.__buffer = buf
+                self.__buffer_pos = chunk_start + needed
+                self.__position -= len(self.__buffer) - self.__buffer_pos
             else:
                 data.append(chunk_data)
-
-        self.__buffer = extra
-        self.__position -= len(extra)
+            received += len(chunk_data)
 
         # Detect extra chunks after reading the entire file.
-        if size == remainder and self.__chunk_iter and not extra:
+        if size == remainder and self.__chunk_iter:
             try:
                 self.__chunk_iter.next()
             except StopIteration:
@@ -637,6 +650,7 @@ class GridOut(io.IOBase):
 
         self.__position = new_pos
         self.__buffer = EMPTY
+        self.__buffer_pos = 0
         if self.__chunk_iter:
             self.__chunk_iter.close()
             self.__chunk_iter = None