@ShaneHarvey 2022-11-07
Note that my first approach was to use a memoryview() for the buffered chunk (`self.__buffer`) but this was cumbersome for a few reasons:

1. The readchunk() method needs to return a bytes, not a memoryview, so we would need to pay the cost of copying to bytes anyway.
2. The readline() method uses bytes.index() to find the next newline which memoryview does not support so it's handy to keep the bytes around.